+++ src/MangoFaaS.Firecracker.Node/Pooling/IFirecrackerProcessPool.cs
@@ -1,0 +1,8 @@
+using MangoFaaS.Firecracker.Node.Pooling;
+
+namespace MangoFaaS.Firecracker.Node.Pooling;
+
+public interface IFirecrackerProcessPool
+{
+    Task<FirecrackerLease> AcquireAsync(string functionHash, CancellationToken cancellationToken = default);
+}
